+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@
 * #46 Test API: smart DB cleanup - handle dependent tables
 * #47 Test API: an option to clear Cayenne caches after deleting test data
 * #48 CayenneTestModule - replace 'contributeSchemaListeners' with an extender
+* #49 Upgrade to Cayenne 4.0.B2
 
 ## 0.20